Correct MinesweeperMaster solution in C#
This uses a depth-first search and assumes that the initial click point
is the top-left hand corner. It ran very quickly, even on the large
problem set, so it might be worth iterating through all points in the
top left quadrant as the initial click point.

This took me around 2 hours to code in C# (and I already knew what I
wanted to do, as it involved a few improvements to the Scala algorithm
that I couldn't get debugged during the live round).

Although C# is more verbose than Scala, the Visual Studio code
completion and debugger is streets ahead of the Eclipse IDE (although I
hardly needed the debugger).