Refactor and debug Scala code for MinesweepMaster
Refactor the Scala algorithm to:
1. Use a similar improved algorithm to the C# solution
2. Not use a SearchNode class. Instead just call a recursive search()
method.

There were a number of bugs to fix, mostly around confusing row & column
variables with x and y parameters.

The output files have been accepted by the Google Code Jam adjudication
site.

The algorithm ran in almost 1 second on the large problem set, compared
to 3.6 seconds for the C# solution. However this difference may just be
due to the frontier nodes being expanded in a different sequence in the
Scala solution compared to the C# solution. This still indicates that
the immutable data structures in the Scala solution compare favourably
with the mutable data structures used in the C# solution.

The C# solution had around 70% more lines of code than the Scala
solution, but was quicker to write and virtually worked first time. This
is partially because I used a 2-D array for the C# solution, but an
immutable Vector data type for the Scala solution. Some of my Scala bugs
were from the translation from 1-D to 2-D coordinates.