C# failed submission: problem B in 2014 round 1B
I completed this at around the 2h40 mark (10 minutes after the end of
competition).
I'm not sure if it is passing any problem sets yet (except for the
test.in file, which it does pass).

Small attempt 0 was downloaded before the end of the competition.
Small attempt 1 was downloaded shortly after.